remove references to synthetic population spec
