+++ src/main/kotlin/org/wfanet/measurement/integration/common/InProcessEdpSimulator.kt
@@ -54,7 +54,6 @@ import org.wfanet.measurement.eventdataprovider.privacybudgetmanagement.testing.
 import org.wfanet.measurement.eventdataprovider.requisition.v2alpha.common.InMemoryVidIndexMap
 import org.wfanet.measurement.loadtest.dataprovider.EdpSimulator
 import org.wfanet.measurement.loadtest.dataprovider.SyntheticGeneratorEventQuery
-import org.wfanet.measurement.loadtest.dataprovider.toPopulationSpecWithoutAttributes
 
 /** An in process EDP simulator. */
 class InProcessEdpSimulator(
@@ -89,8 +88,7 @@ class InProcessEdpSimulator(
   private val delegate: EdpSimulator
 
   init {
-    val populationSpec: PopulationSpec =
-      eventQuery.populationSpec.toPopulationSpecWithoutAttributes()
+    val populationSpec: PopulationSpec = eventQuery.populationSpec
     val vidIndexMap = InMemoryVidIndexMap.build(populationSpec)
 
     delegate =
