+++ src/main/kotlin/org/wfanet/measurement/integration/common/SyntheticGenerationSpecs.kt
@@ -17,9 +17,11 @@
 package org.wfanet.measurement.integration.common
 
 import com.google.protobuf.Message
+import com.google.protobuf.TypeRegistry
 import java.nio.file.Paths
 import org.wfanet.measurement.api.v2alpha.PopulationSpec
 import org.wfanet.measurement.api.v2alpha.event_group_metadata.testing.SyntheticEventGroupSpec
+import org.wfanet.measurement.api.v2alpha.event_templates.testing.Person
 import org.wfanet.measurement.common.getRuntimePath
 import org.wfanet.measurement.common.parseTextProto
 
@@ -29,12 +31,19 @@ object SyntheticGenerationSpecs {
     Paths.get("wfa_measurement_system", "src", "main", "k8s", "testing", "data")
   private val TEST_DATA_RUNTIME_PATH = getRuntimePath(TEST_DATA_PATH)!!
 
+  /**
+   * Type registry used when parsing the v2alpha [PopulationSpec] textprotos in this directory.
+   *
+   * The textprotos embed event template attribute messages (today only [Person]) inside
+   * `google.protobuf.Any` packed values, so the parser needs the descriptors for those template
+   * types.
+   */
+  private val TYPE_REGISTRY: TypeRegistry =
+    TypeRegistry.newBuilder().add(Person.getDescriptor()).build()
+
   /** v2alpha [PopulationSpec] for synthetic generation of population ~34,000,000. */
   val POPULATION_SPEC_LARGE: PopulationSpec by lazy {
-    loadTestData(
-      "synthetic_population_spec_large.textproto",
-      PopulationSpec.getDefaultInstance(),
-    )
+    loadTestData("synthetic_population_spec_large.textproto", PopulationSpec.getDefaultInstance())
   }
 
   /** EventGroup specs for synthetic generation based on [POPULATION_SPEC_LARGE]. */
@@ -75,10 +84,7 @@ object SyntheticGenerationSpecs {
 
   /** v2alpha [PopulationSpec] for synthetic generation of population ~100,000. */
   val POPULATION_SPEC_SMALL: PopulationSpec by lazy {
-    loadTestData(
-      "synthetic_population_spec_small.textproto",
-      PopulationSpec.getDefaultInstance(),
-    )
+    loadTestData("synthetic_population_spec_small.textproto", PopulationSpec.getDefaultInstance())
   }
 
   /** EventGroup specs for synthetic generation based on [POPULATION_SPEC_SMALL]. */
@@ -114,6 +120,10 @@ object SyntheticGenerationSpecs {
   }
 
   private fun <T : Message> loadTestData(fileName: String, defaultInstance: T): T {
-    return parseTextProto(TEST_DATA_RUNTIME_PATH.resolve(fileName).toFile(), defaultInstance)
+    return parseTextProto(
+      TEST_DATA_RUNTIME_PATH.resolve(fileName).toFile(),
+      defaultInstance,
+      TYPE_REGISTRY,
+    )
   }
 }
