+++ src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider/BUILD.bazel
@@ -292,15 +292,3 @@ kt_jvm_library(
     ],
 )
 
-kt_jvm_library(
-    name = "population_spec_converter",
-    srcs = [
-        "PopulationSpecConverter.kt",
-    ],
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:event_templates",
-        "//src/main/proto/wfa/measurement/api/v2alpha:population_spec_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha/event_group_metadata/testing:simulator_synthetic_data_spec_kt_jvm_proto",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
-    ],
-)
