+++ src/test/kotlin/org/wfanet/measurement/loadtest/dataprovider/BUILD.bazel
@@ -111,32 +111,11 @@ java_test(
 )
 
 kt_jvm_test(
-    name = "SyntheticDataGenerationTest",
-    srcs = ["SyntheticDataGenerationTest.kt"],
+    name = "SyntheticDataGenerationPopulationSpecTest",
+    srcs = ["SyntheticDataGenerationPopulationSpecTest.kt"],
     data = [
         "//src/main/proto/wfa/measurement/loadtest/dataprovider:specs",
     ],
-    test_class = "org.wfanet.measurement.loadtest.dataprovider.SyntheticDataGenerationTest",
-    deps = [
-        "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:synthetic_data_generation",
-        "//src/main/proto/wfa/measurement/api/v2alpha/event_group_metadata/testing:simulator_synthetic_data_spec_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/api/v2alpha/event_templates/testing:test_event_kt_jvm_proto",
-        "@wfa_common_jvm//imports/java/com/google/common/truth",
-        "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
-        "@wfa_common_jvm//imports/java/com/google/protobuf",
-        "@wfa_common_jvm//imports/java/org/junit",
-        "@wfa_common_jvm//imports/kotlin/kotlin/test",
-        "@wfa_common_jvm//imports/kotlin/org/mockito/kotlin",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/identity",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/testing",
-        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/storage/filesystem:client",
-    ],
-)
-
-kt_jvm_test(
-    name = "SyntheticDataGenerationPopulationSpecTest",
-    srcs = ["SyntheticDataGenerationPopulationSpecTest.kt"],
     test_class = "org.wfanet.measurement.loadtest.dataprovider.SyntheticDataGenerationPopulationSpecTest",
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider:synthetic_data_generation",
